refactor : validation 추가
