+++ src/main/java/com/example/delivery_app/domain/menu/exception/ErrorCode.java
@@ -1,22 +1,24 @@
 package com.example.delivery_app.domain.menu.exception;
 
+import org.springframework.http.HttpStatus;
+
+import com.example.delivery_app.common.dto.ResponseCode;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum ErrorCode {
+public enum ErrorCode implements ResponseCode {
 
 	// Menu 관련
-	MENU_NOT_FOUND(404, "Not Found", "M001", "찾는 메뉴가 없습니다."),
-	MISMATCH_ERROR(400, "Bad Request", "M002", "가게와 메뉴가 일치하지 않습니다."),
+	MENU_NOT_FOUND(HttpStatus.NOT_FOUND, "찾는 메뉴가 없습니다."),
+	MISMATCH_ERROR(HttpStatus.BAD_REQUEST, "가게와 메뉴가 일치하지 않습니다."),
 
 	// MenuOption 관련
-	MENU_OPTION_NOT_FOUND(404, "Not Found", "O001", "찾는 옵션이 없습니다."),
-	MENU_ALREADY_DELETED(400, "Bad Request", "O002", "이미 삭제된 옵션입니다.");
+	MENU_OPTION_NOT_FOUND(HttpStatus.NOT_FOUND, "찾는 옵션이 없습니다."),
+	MENU_ALREADY_DELETED(HttpStatus.BAD_REQUEST, "이미 삭제된 옵션입니다.");
 
-	private final int status;
-	private final String error;
-	private final String code;
+	private final HttpStatus httpStatus;
 	private final String message;
 }
